requirements.txt

ipython
jupyter
sqlalchemy
us
boltons
invoke
cached-property
tqdm
pandas
attrs
pytest
PyYAML
